himitsu: force prod over i2p and enable mainnet
The prod release will now allow rpc to connect over i2p
only. This is for a future effort of public rpc threads
over i2p with some form of alternate authentication, such
as the prokurilo reverse proxy authentication server. It
could be used as is but opens up an attack vector of protecting
.b32.i2p address just as you would a private key. Mainnet
restrictions on transfers and contacts are removed.