analitiko: kmeans: set dynamic outlier threshold
Update pacakge installation to be controlled by a DEV_ENV
environment variable. Use "local" in the .Renvrion file to
avoid running the installs on each build. Use "docker"
when baking a new image. Model outlier threshold is set
dynamically to the bounds of the second standard deviation of
transaction size.